fixed: moving autocompletion routine to a dedicated .rs file
